Divide user registration into 2 forms
