+++ src/pages/authentication/Login/Login.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '../../../components/forms/LoginForm';
+import LoginForm from '../../../components/forms/auth/LoginForm';
 import { TransparentBoxStyled } from '../../../components/Layout/Box.styled';
 
 const Login: React.FC = () => (
